requirements.txt
fastapi
uvicorn
image-quality
requests
pillow
